backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Async DB
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Task queue
celery==5.4.0
redis==5.2.1

# Settings
pydantic==2.10.4
pydantic-settings==2.7.0

# LLM Providers
anthropic==0.42.0
google-generativeai==0.8.3
openai==1.59.4

# Audio / ASR
openai-whisper==20240930

# Document parsing
python-pptx==1.0.2
pymupdf==1.25.1          # PDF extraction (fitz)
Pillow==11.1.0

# Embeddings / alignment
numpy==2.2.1
scipy==1.15.0

# HTTP client
httpx==0.28.1

# Dev / test
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
